% 
% Copyright 2001 Mayo Foundation for Medical Education and Research. 
% 
% This program is free software; you can redistribute it and/or
% modify it under the terms of the GNU General Public License
% as published by the Free Software Foundation; either version 2
% of the License, or (at your option) any later version.
% 
% This program is distributed in the hope that it will be useful,
% but WITHOUT ANY WARRANTY; without even the implied warranty of
% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
% GNU General Public License for more details.
% 
% You should have received a copy of the GNU General Public License
% along with this program; if not, write to the Free Software
% Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  
% 02111-1307, USA.
% 
% 
\name{banner}
\alias{banner}
\title{
  Print a nice banner
}
\description{

}
\usage{
banner(str, banner.width=80, char.perline=60, border="=")
}
\arguments{


\item{str}{
character string - a title within the banner
}
\item{banner.width }{
width of banner
}
\item{char.perline }{
number of characters per line for the title
}
\item{border}{
type of character for the border
}
}
\value{

}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{

banner("This is a pretty banner", banner.width=40, char.perline=30)

# ========================================
#          This is a pretty banner
# ========================================
}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
